Bug 1362872 - Use mbcs encoding for vswhere output. r=gps
We add vswhere.exe to support VS2017.  But the description output of this command is localized, so it causes UnicodeDecodeError.

So we should use 'mbcs' encoding for this output on Windows platform. And for minor languages, we should also use replace to avoid decode error.

test_toolchain_configure.py calls vc_compiler_path without correct host.kernel, so we should check current python platform.


MozReview-Commit-ID: AkryAzrgSzs