Bug 1350152 - Don't rely on gBrowser in nsLoginManagerPrompter.js. r=mkaply a=jorgk
In bug 1266836 we fixed this code to properly work in e10s,
but made the mistake of using gBrowser in toolkit code, breaking
XUL apps that don't have a gBrowser. This solution implements
a different code-path for non-e10s and returns a chrome
window for consumers that don't offer a browser.

MozReview-Commit-ID: 929ahDenzu5